requirements.txt
langchain==0.0.335
openai==1.2.4
streamlit==1.28.2
pinecone-client==2.2.4
pandas==2.0.3
tiktoken==0.5.1